featherify-main/server/requirements.txt
Django>=3.1.0,<3.2.0
djangorestframework>=3.12.0,<3.13.0
Pillow>=8.1.0,<8.2.0
django-cors-headers>=3.7.0,<3.8.0
urllib3>=1.26.3,<1.27.0